Вынес функции аккаунтов и прокси в отдельные файлы
